Move views to pages folder
