+++ front/src/pages/DayView/DayView.module.scss
@@ -1,4 +1,4 @@
-@import '../shared.module.scss';
+@import '../../shared.module.scss';
 
 .sevenGrid {
   display: flex;
