+++ front/src/pages/DayView/DayView.tsx
@@ -8,11 +8,11 @@ import CircularProgress from '@mui/material/CircularProgress';
 
 // Moment for date handling
 import moment from 'moment';
-import { JumpToCurrent } from '../utils/Utils';
+import { JumpToCurrent } from '../../utils/Utils';
 import info from '@/assets/Info.png';
-import api from '../api/api';
-import css from './Dayview.module.scss';
-import { DeviceStatusList } from '../DeviceStatusList/DeviceStatusList';
+import api from '../../api/api';
+import css from './DayView.module.scss';
+import { DeviceStatusList } from '../../DeviceStatusList/DeviceStatusList';
 import { ViewChanger } from '@/lib/components/ViewChanger';
 import { useQuery } from 'react-query';
 import { DateHeader } from '@/lib/components/DateHeader';
