+++ front/src/pages/MonthView/MonthView.module.scss
@@ -1,4 +1,4 @@
-@import '../shared.module.scss';
+@import '../../shared.module.scss';
 
 .hoverHand:hover {
   cursor: pointer;
