+++ front/src/pages/WeekView/WeekView.module.scss
@@ -1,4 +1,4 @@
-@import '../shared.module.scss';
+@import '../../shared.module.scss';
 
 .infoBox {
   display: flex;
