+++ front/src/pages/WeekView/WeekView.test.tsx
@@ -1,9 +1,9 @@
 import { waitFor, render, screen } from '@testing-library/react';
 import { HashRouter as Router } from 'react-router-dom';
 import { act } from 'react-dom/test-utils';
-import { getSchedulingWeek } from '../utils/Utils';
-import { Weekview } from './Weekview';
-import testUtils, { schedulingWeek} from '../_TestUtils/TestUtils';
+import { getSchedulingWeek } from '../../utils/Utils';
+import { Weekview } from './WeekView';
+import testUtils, { schedulingWeek} from '../../_TestUtils/TestUtils';
 import { TestProviders } from '@/_TestUtils/TestProvides';
 
 const { date, week } = testUtils;
@@ -13,7 +13,7 @@ vi.mock('@/lib/components/InfoBox', () => ({
   InfoBox: () => <div data-testid="mockInfoBox">Mock InfoBox</div>,
 }));
 
-vi.mock(import("../utils/Utils"), async (importOriginal) => {
+vi.mock(import("../../utils/Utils"), async (importOriginal) => {
   const actual = await importOriginal()
   return {
     ...actual,
